requirements.txt
channels-redis==3.4.1
Django==4.1.1
djangochannelsrestframework==1.1.0
djangorestframework==3.13.1
